requirements.txt
pandas
numpy
streamlit
matplotlib
plotly
google-generativeai
openai